pyproject.toml
[tool.poetry]
name = "schedule_bot"
version = "0.1.0"
description = ""
authors = ["rdfsx <m315498@gmail.com>"]
packages = [{ include = "app" }]

[tool.poetry.dependencies]
python = "^3.8"
aiogram = "^2.14.3"
environs = "^9.3.3"
loguru = "^0.5.3"
gino = "^1.0.1"
alembic = "^1.7.1"
psycopg2-binary = "^2.9.1"
python-dotenv = "^0.19.0"
beautifulsoup4 = "^4.9.3"
APScheduler = "^3.7.0"
chatbase = {git = "https://github.com/google/chatbase-python.git"}
aiomysql = "^0.0.21"
aiojobs = "^0.3.0"
requests = "^2.26.0"
aioredis = "1.3.1"
SQLAlchemy = "~=1.3.22"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
